Refactor: 커뮤니티, 이벤트 게시판 게시글 작성 폼 컴포넌트 분리
