+++ client/src/components/communityAndEvent/PostWriteForm.tsx
@@ -1,30 +1,21 @@
 import React, { ChangeEvent, useEffect, useRef, useState } from "react";
 import ImageUploader from "./ImageUploader";
-import NewTags from "./NewTags";
 import NewTagForm from "./NewTagForm";
-import { ICommunityDetail } from "../../models/community.model";
-import { IEventDetail } from "../../models/event.model";
+import NewTags from "./NewTags";
 
-type Post = ICommunityDetail | IEventDetail;
-
-const isClosedType = (post: Post): post is IEventDetail => "isClosed" in post;
-interface IProps<T> {
+interface IProps {
   boardCategory: "community" | "event";
-  post: T;
-  editPost: (formData: FormData) => void;
+  addPost: (formData: FormData) => void;
 }
 
-const PostEditForm = <T extends Post>({ boardCategory, post, editPost }: IProps<T>) => {
-  const oldImages = post.images.map((image) => image.url);
-  const oldTags = post.tags.map((tag) => tag.tag);
-  const textareaRef = useRef<HTMLTextAreaElement>(null);
+const PostWriteForm = ({ boardCategory, addPost }: IProps) => {
+  const textareaRef = useRef<HTMLTextAreaElement | null>(null);
+  const [title, setTitle] = useState("");
+  const [content, setContent] = useState("");
   const [isOpen, setIsOpen] = useState(false);
-  const [title, setTitle] = useState(post.title);
-  const [content, setContent] = useState(post.content);
-  const [newTags, setNewTags] = useState<string[]>([...oldTags]);
-  const [newImages, setNewImages] = useState<(string | File)[]>(oldImages);
+  const [newTags, setNewTags] = useState<string[]>([]);
   const [isClosed, setIsClosed] = useState(false);
-  const [tagsToDelete, setTagsToDelete] = useState<string[]>([]);
+  const [newImages, setNewImages] = useState<(string | File)[]>([]);
 
   const handleResizeHeight = () => {
     if (textareaRef.current) {
@@ -46,13 +37,7 @@ const PostEditForm = <T extends Post>({ boardCategory, post, editPost }: IProps<
     }
   };
 
-  const removeTags = (indexToRemove: number, tag?: string) => {
-    const foundTag = newTags.find((item) => item === tag);
-
-    if (foundTag) {
-      setTagsToDelete([...tagsToDelete, foundTag]);
-    }
-
+  const removeTags = (indexToRemove: number) => {
     const filter = newTags.filter((_, index) => index !== indexToRemove);
     setNewTags(filter);
   };
@@ -69,42 +54,21 @@ const PostEditForm = <T extends Post>({ boardCategory, post, editPost }: IProps<
     setNewImages([...images]);
   };
 
-  // 삭제할 태그들을 저장합니다.
-  const addTagsToDelete = (tags: string[]) => {
-    setTagsToDelete([...tags]);
-  };
-
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-
-    //NOTE 이게 맞나..
-
-    // DB에서 삭제할 태그 ID만 뽑아내기
-    const filteredTagIds = post.tags.filter((tag) => tagsToDelete.includes(tag.tag)).map((item) => item.tagId);
-
-    // 새로운 태그 중에 기존 태그에 없는 것만 골라내기
-    const filteredTags = newTags.filter((tag) => !oldTags.includes(tag));
-
-    // DB에서 삭제할 이미지 ID만 뽑아내기
-    const filteredImageIds = post.images.filter((image) => !newImages.includes(image.url)).map((item) => item.imageId);
-
-    // DB에 저장할 이미지만 골라내기
-    const filteredImages = newImages.filter((image): image is File => image instanceof File);
 
     const formData = new FormData();
 
     formData.append("title", title);
     formData.append("content", content);
     boardCategory === "event" && formData.append("isClosed", isClosed ? "true" : "");
-    formData.append("deleteTagIds", JSON.stringify(filteredTagIds) ?? []);
-    formData.append("tags", JSON.stringify(filteredTags) ?? []);
-    formData.append("deleteImageIds", JSON.stringify(filteredImageIds) ?? []);
+    formData.append("tags", JSON.stringify(newTags));
 
-    filteredImages.forEach((image) => {
+    Array.from(newImages).forEach((image) => {
       formData.append("images", image);
     });
 
-    editPost(formData);
+    addPost(formData);
   };
 
   useEffect(() => {
@@ -139,7 +103,7 @@ const PostEditForm = <T extends Post>({ boardCategory, post, editPost }: IProps<
 
           <ImageUploader newImages={newImages} setNewImageFiles={setNewImageFiles} />
 
-          {isClosedType(post) && (
+          {boardCategory === "event" && (
             <div className="status-selection">
               <span className="status-title">모집</span>
               <div className="status-buttons">
@@ -179,16 +143,9 @@ const PostEditForm = <T extends Post>({ boardCategory, post, editPost }: IProps<
         <button className="post-submit">수정완료</button>
       </form>
 
-      {isOpen && (
-        <NewTagForm
-          initialTags={newTags}
-          addTagsToDelete={addTagsToDelete}
-          addNewTags={addNewTags}
-          handleTagFormOpen={handleTagFormOpen}
-        />
-      )}
+      {isOpen && <NewTagForm initialTags={newTags} addNewTags={addNewTags} handleTagFormOpen={handleTagFormOpen} />}
     </div>
   );
 };
 
-export default PostEditForm;
+export default PostWriteForm;
