+++ client/src/styles/scss/components/communityAndEvent/postList.scss
@@ -1,4 +1,5 @@
 @import "../../base/common.scss";
 
 .list {
+  min-height: 100vh;
 }
