+++ client/src/styles/scss/pages/community/communityPostWrite.scss
@@ -1,13 +1,14 @@
 @import "../../base/common.scss";
 
 .community-post-write {
-  height: 100vh;
+  min-height: 100vh;
+
   .post-form {
     display: flex;
     flex-direction: column;
     justify-content: space-between;
-    height: calc(100vh - 60rem);
     padding-top: 41rem;
+    min-height: calc(100vh - 60rem);
 
     button {
       background-color: transparent;
@@ -29,7 +30,7 @@
       }
     }
 
-    textarea {
+    .post-textarea {
       width: 100%;
       height: 19rem;
       font-family: "medium";
@@ -62,8 +63,7 @@
     .post-submit {
       @include bigButton("on");
       font-family: "bold";
-      margin: 0 20rem;
-      margin-bottom: 20rem; // 14rem은 너무 좁아서 수정
+      margin: 20rem;
     }
   }
